moved arrays tests to its own source file
